requirements.txt
sentence-transformers==2.6.0
flask
langchain
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain_huggingface
-e .